+++ Tests/src/PublicAPITests/OnlineRealtimeEngineTests.cpp
@@ -672,3 +672,100 @@ CSP_PUBLIC_TEST_WITH_MOCKS(CSPEngine, OnlineRealtimeEngineTests, ConstructWithCo
 
     EXPECT_NE(Engine.GetComponentSchemaRegistry()->Find(ExampleSchemaId), nullptr);
 }
+
+CSP_PUBLIC_TEST_WITH_MOCKS(CSPEngine, OnlineRealtimeEngineTests, ConstructFromJsonWithComponentSchemas)
+{
+    auto& SystemsManager = csp::systems::SystemsManager::Get();
+
+    const auto AllPropertyTypesSchemaId = ComponentSchema::TypeIdType { 123 };
+    const auto InvalidSchemaId = ComponentSchema::TypeIdType { 789 };
+    const auto EmptySchemaId = ComponentSchema::TypeIdType { 456 };
+
+    constexpr auto RawJsonFirstFile = R"([
+        {
+            "typeId": 123,
+            "name": "AllPropertyTypesComponent",
+            "properties": [
+                {
+                    "key": 0,
+                    "name": "stringProperty",
+                    "type": "string",
+                    "defaultValue": "hello"
+                },
+                {
+                    "key": 1,
+                    "name": "floatProperty",
+                    "type": "float",
+                    "defaultValue": 1.5
+                },
+                {
+                    "key": 2,
+                    "name": "intProperty",
+                    "type": "int",
+                    "defaultValue": 42
+                },
+                {
+                    "key": 3,
+                    "name": "boolProperty",
+                    "type": "bool",
+                    "defaultValue": true
+                },
+                {
+                    "key": 4,
+                    "name": "vec2Property",
+                    "type": "vec2",
+                    "defaultValue": [1.0, 2.0]
+                },
+                {
+                    "key": 5,
+                    "name": "vec3Property",
+                    "type": "vec3",
+                    "defaultValue": [1.0, 2.0, 3.0]
+                },
+                {
+                    "key": 6,
+                    "name": "vec4Property",
+                    "type": "vec4",
+                    "defaultValue": [1.0, 2.0, 3.0, 4.0]
+                }
+            ]
+        }
+    ])";
+
+    constexpr auto RawJsonThirdFile = R"([
+        {
+            "typeId": 789,
+            "name": "InvalidComponent",
+            "properties": [
+                {
+                    "key": 0,
+                    "name": "missingTypeProperty",
+                    "defaultValue": "hello"
+                }
+            ]
+        },
+        {
+            "typeId": 456,
+            "name": "EmptyComponent",
+            "properties": []
+        }
+    ])";
+
+    const auto JsonSchemas = csp::common::List<csp::common::String> {
+        csp::common::String { RawJsonFirstFile },
+        csp::common::String { "this is not json at all" },
+        csp::common::String { RawJsonThirdFile },
+    };
+
+    const auto Engine = OnlineRealtimeEngine {
+        *SystemsManager.GetMultiplayerConnection(),
+        *SystemsManager.GetLogSystem(),
+        *SystemsManager.GetEventBus(),
+        *SystemsManager.GetScriptSystem(),
+        JsonSchemas,
+    };
+
+    EXPECT_NE(Engine.GetComponentSchemaRegistry()->Find(AllPropertyTypesSchemaId), nullptr);
+    EXPECT_NE(Engine.GetComponentSchemaRegistry()->Find(EmptySchemaId), nullptr);
+    EXPECT_EQ(Engine.GetComponentSchemaRegistry()->Find(InvalidSchemaId), nullptr);
+}